refactor(#267 F1): split 1321-LOC tts/ssml.rs into ssml/{segment,rate,walker,warnings,mod}.rs
The audit (#267 P0.F1): `rust/src/tts/ssml.rs` was 1321 LOC holding the
`Segment` enum, three free rate-parse helpers, the 233-LOC `parse()`
glue, a 117-LOC recursive walker, warn-bucket constants, a doctype
check, and a 660-LOC test block. The hottest file in the tree (#236
prosody, #232 abbrev, #122 SSML) — every new tag dropped into one
mailbox.

Split:

- `tts/ssml/mod.rs` (~320 LOC + 660 LOC tests): façade with re-exports,
  `pub fn parse()`, and the three small in-loop helpers (`tag_name`,
  `contains_doctype`, `extract_inner_text`) that close over the
  `ssml_parser` enum.
- `tts/ssml/segment.rs` (38 LOC): the `Segment` enum + `DEFAULT_BREAK`.
- `tts/ssml/rate.rs` (159 LOC): `parse_rate_value`, `find_relative_rate`,
  `has_structural_source_siblings` — everything that decides what a
  `<prosody rate>` value means in the input string.
- `tts/ssml/walker.rs` (163 LOC): `parse_inner_spans` (the recursive
  walker for whole-utterance prosody bodies) + its private helpers
  `span_priority` and `push_text_slice`.
- `tts/ssml/warnings.rs` (8 LOC): `WARN_PROSODY_*` bucket keys.

Public surface unchanged — `tts::ssml::{parse, Segment}` still resolve
the same way (`pub use segment::Segment` in mod.rs). Tests stay inside
`mod.rs::tests` for this PR; moving the integration-shaped `<speak>`
scenarios to `rust/tests/ssml_integration.rs` is tracked as F8 in #267.

Gates:
  cargo fmt --check                            clean
  cargo clippy --all-targets -- -D warnings    clean (default features)
  cargo test --lib tts::ssml                   45 / 45 passing

Refs #267
